# CI: Speed up Azure tests