--- a/mne/io/fiff/tests/test_raw_fiff.py
+++ b/mne/io/fiff/tests/test_raw_fiff.py
@@ -382,14 +382,15 @@ def test_concatenate_raws(on_mismatch):
 @testing.requires_testing_data
 @pytest.mark.parametrize('mod', (
     'meg',
-    pytest.param('raw', marks=[pytest.mark.filterwarnings(
-        'ignore:.*naming conventions.*:RuntimeWarning')]),
+    pytest.param('raw', marks=[
+        pytest.mark.filterwarnings(
+            'ignore:.*naming conventions.*:RuntimeWarning'),
+        pytest.mark.slowtest]),
 ))
 def test_split_files(tmpdir, mod, monkeypatch):
     """Test writing and reading of split raw files."""
     raw_1 = read_raw_fif(fif_fname, preload=True)
     # Test a very close corner case
-    raw_crop = raw_1.copy().crop(0, 1.)
 
     assert_allclose(raw_1.buffer_size_sec, 10., atol=1e-2)  # samp rate
     split_fname = tmpdir.join(f'split_raw_{mod}.fif')
@@ -1278,7 +1279,7 @@ def test_raw_copy():
 def test_to_data_frame():
     """Test raw Pandas exporter."""
     from pandas import Timedelta
-    raw = read_raw_fif(test_fif_fname, preload=True)
+    raw = read_raw_fif(test_fif_fname).crop(0, 1).load_data()
     _, times = raw[0, :10]
     df = raw.to_data_frame(index='time')
     assert ((df.columns == raw.ch_names).all())
@@ -1683,15 +1684,18 @@ def test_memmap(tmpdir):
 
 
 # These are slow on Azure Windows so let's do a subset
-@pytest.mark.parametrize('split', [
-    False,
-    pytest.param(True, marks=pytest.mark.slowtest),
-])
 @pytest.mark.parametrize('kind', [
     'file',
     pytest.param('bytes', marks=pytest.mark.slowtest),
 ])
-@pytest.mark.parametrize('preload', (True, str))
+@pytest.mark.parametrize('preload', [
+    True,
+    pytest.param(str, marks=pytest.mark.slowtest),
+])
+@pytest.mark.parametrize('split', [
+    False,
+    pytest.param(True, marks=pytest.mark.slowtest),
+])
 def test_file_like(kind, preload, split, tmpdir):
     """Test handling with file-like objects."""
     if split: