--- a/mne/minimum_norm/tests/test_inverse.py
+++ b/mne/minimum_norm/tests/test_inverse.py
@@ -309,8 +309,8 @@ def test_inverse_operator_channel_ordering(evoked, noise_cov):
     ('MNE', 89, 92, dict(limit_depth_chs='whiten')),  # sparse default
     ('dSPM', 96, 98, 0.8),
     ('sLORETA', 100, 100, 0.8),
-    ('eLORETA', 100, 100, None),
-    ('eLORETA', 100, 100, 0.8),
+    pytest.param('eLORETA', 100, 100, None, marks=pytest.mark.slowtest),
+    pytest.param('eLORETA', 100, 100, 0.8, marks=pytest.mark.slowtest),
 ])
 def test_localization_bias_fixed(bias_params_fixed, method, lower, upper,
                                  depth):
@@ -333,8 +333,8 @@ def test_localization_bias_fixed(bias_params_fixed, method, lower, upper,
     ('dSPM', 85, 87, 0.8, 0.2),
     ('sLORETA', 100, 100, 0.8, 0.2),
     ('eLORETA', 99, 100, None, 0.2),
-    ('eLORETA', 99, 100, 0.8, 0.2),
-    ('eLORETA', 99, 100, 0.8, 0.001),
+    pytest.param('eLORETA', 99, 100, 0.8, 0.2, marks=pytest.mark.slowtest),
+    pytest.param('eLORETA', 99, 100, 0.8, 0.001, marks=pytest.mark.slowtest),
 ])
 @pytest.mark.parametrize('pick_ori', (None, 'vector'))
 def test_localization_bias_loose(bias_params_fixed, method, lower, upper,
@@ -376,11 +376,15 @@ def test_localization_bias_loose(bias_params_fixed, method, lower, upper,
          dict(limit_depth_chs='whiten'), 1),  # sparse default
         ('dSPM', 40, 45, 0.96, 0.97, {}, 0.8, 1),
         ('sLORETA', 93, 95, 0.95, 0.96, {}, 0.8, 1),
-        ('eLORETA', 93, 100, 0.95, 0.96,
-         dict(method_params=dict(force_equal=True)), None, 1),
-        ('eLORETA', 100, 100, 0.98, 0.99, {}, None, 1.0),
-        ('eLORETA', 100, 100, 0.98, 0.99, {}, 0.8, 1.0),
-        ('eLORETA', 100, 100, 0.98, 0.99, {}, 0.8, 0.999),
+        pytest.param('eLORETA', 93, 100, 0.95, 0.96,
+                     dict(method_params=dict(force_equal=True)), None, 1,
+                     marks = pytest.mark.slowtest),
+        pytest.param('eLORETA', 100, 100, 0.98, 0.99, {}, None, 1.0,
+                     marks=pytest.mark.slowtest),
+        pytest.param('eLORETA', 100, 100, 0.98, 0.99, {}, 0.8, 1.0,
+                     marks=pytest.mark.slowtest),
+        pytest.param('eLORETA', 100, 100, 0.98, 0.99, {}, 0.8, 0.999,
+                     marks=pytest.mark.slowtest),
     ]
 )
 def test_localization_bias_free(bias_params_free, method, lower, upper,
@@ -624,6 +628,8 @@ def assert_var_exp_log(log, lower, upper):
 @pytest.mark.parametrize('pick_ori', (None, 'vector'))
 def test_inverse_residual(evoked, method, pick_ori):
     """Test MNE inverse application."""
+    if method == 'eLORETA' and pick_ori == 'vector':  # works but slow
+        return
     # use fname_inv as it will be faster than fname_full (fewer verts and chs)
     evoked = evoked.pick_types(meg=True)
     if pick_ori is None:  # use fixed
@@ -657,6 +663,7 @@ def test_inverse_residual(evoked, method, pick_ori):
         assert_array_less(np.abs(residual.data), 1e-15)
 
 
+@pytest.mark.slowtest
 def test_make_inverse_operator_fixed(evoked, noise_cov):
     """Test MNE inverse computation (fixed orientation)."""
     fwd = read_forward_solution_meg(fname_fwd)
@@ -987,6 +994,7 @@ def test_apply_mne_inverse_fixed_raw():
     assert_array_almost_equal(stc.data, stc3.data)
 
 
+@pytest.mark.slowtest
 @testing.requires_testing_data
 def test_apply_mne_inverse_epochs():
     """Test MNE with precomputed inverse operator on Epochs."""