In https://dev.azure.com/mne-tools/mne-python/_build/results?buildId=16390&view=logs&jobId=c1ec3add-0fcd-52ea-530a-328457bbfb2e&j=2bd7b19d-6351-5e7f-8417-63f327ab45bc&t=0c49e006-47ab-5675-a040-5d301cde7356 we see:
```
106.95s total  mne/minimum_norm/tests/
104.80s total  mne/tests/test_epochs.py
 92.94s total  mne/beamformer/tests/
 83.44s total  mne/io/fiff/
 70.14s total  mne/preprocessing/tests/test_ica.py
 65.67s total  mne/tests/test_rank.py
```
Let's see why, and maybe mark a few more tests as `slowtest`. Even the Azure run that just passed was too close to the 60-minute mark for comfort...