@larsoner 2021-11-02
Okay now we're generally down to less than half:
```
57.17s total   mne/beamformer/tests/
51.15s total   mne/tests/test_epochs.py
50.17s total   mne/preprocessing/tests/test_ica.py
43.19s total   mne/minimum_norm/tests/
26.38s total   mne/io/fiff/
 0.98s total   mne/tests/test_rank.py
```
I'll restore `azure_pipelines.yml` then hopefully this will be good to go